+++ pages/admintasks/actions.php
@@ -44,6 +44,16 @@ For support and installation notes visit http://www.hlxcommunity.com
         die ("Access denied!");
 	}
 
+	// Check if the 'game' parameter is set in the URL
+	if (isset($_GET['game'])) {
+		// Assign the value of 'game' to the variable $gamemode
+		$gamecode = $_GET['game'];
+
+	} else {
+		// 'game' parameter is not set in the URL
+		$gamecode = null;
+	}
+
 	$edlist = new EditList("id", "hlstats_Actions", "game", false);
 	$edlist->columns[] = new EditListColumn("game", "Game", 0, true, "hidden", $gamecode);
 	$edlist->columns[] = new EditListColumn("code", "Action Code", 15, true, "text", "", 64);
@@ -80,7 +90,7 @@ For support and installation notes visit http://www.hlxcommunity.com
 			}
 		?>
 		<p class="ms-4">You can make an action map-specific by prepending the map name and an underscore to the Action Code. For example, if the map "<b>rock2</b>" has an action "<b>goalitem</b>" then you can either make the action code just "<b>goalitem</b>" (in which case it will match all maps) or you can make it "<b>rock2_goalitem</b>" to match only on the "rock2" map.</p>
-		<form method="post" action="<?php echo $g_options['scripturl']; ?>?mode=admin&amp;task=<?php echo $code; ?>#<?php echo $code; ?>">	
+		<form method="post" name="<?php echo $code; ?>form" action="<?php echo $g_options['scripturl']; ?>?mode=admin&amp;game=<?php echo $gamecode; ?>&task=<?php echo $code; ?>#<?php echo $code; ?>">	
 			<?php
 				
 				$result = $db->query("