+++ pages/admintasks/awards_plyrplyractions_victim.php
@@ -44,6 +44,16 @@ For support and installation notes visit http://www.hlxcommunity.com
         die ('Access denied!');
     }
 	
+	// Check if the 'game' parameter is set in the URL
+	if (isset($_GET['game'])) {
+		// Assign the value of 'game' to the variable $gamemode
+		$gamecode = $_GET['game'];
+
+	} else {
+		// 'game' parameter is not set in the URL
+		$gamecode = null;
+	}
+
 	$edlist = new EditList('awardId', 'hlstats_Awards', 'award', false);
 	$edlist->columns[] = new EditListColumn('game', 'Game', 0, true, 'hidden', $gamecode);
 	$edlist->columns[] = new EditListColumn('awardType', 'Type', 0, true, 'hidden', 'V');
@@ -66,7 +76,7 @@ For support and installation notes visit http://www.hlxcommunity.com
 					echo '<div class="alert alert-danger col-8 text-center mx-auto" role="alert"><strong>' . $edlist->error() . '</strong></div>';
 			}
 		?>
-		<form method="post" action="<?php echo $g_options['scripturl']; ?>?mode=admin&amp;task=<?php echo $code; ?>#<?php echo $code; ?>">	
+		<form method="post" name="<?php echo $code; ?>form" action="<?php echo $g_options['scripturl']; ?>?mode=admin&amp;game=<?php echo $gamecode; ?>&task=<?php echo $code; ?>#<?php echo $code; ?>">	
 			<?php
 				$result = $db->query("
 					SELECT