+++ pages/admintasks/newserver.php
@@ -107,7 +107,7 @@ For support and installation notes visit http://www.hlxcommunity.com
 			</div>
 			<p class="ms-4">Enter the address of a server that you want to accept data from.</p>
 			<p class="ms-4">The "Public Address" should be the address you want shown to users. If left blank, it will be generated from the IP Address and Port. If you are using any kind of log relaying utility (i.e. hlstats.pl will not be receiving data directly from the game servers), you will want to set the IP Address and Port to the address of the log relay program, and set the Public Address to the real address of the game server. You will need a separate log relay for each game server. You can specify a hostname (or anything at all) in the Public Address.</p>
-			<form method="post" action="<?php echo $g_options['scripturl']; ?>?mode=admin&amp;task=<?php echo $code; ?>#<?php echo $code; ?>">
+			<form method="post" name="<?php echo $code; ?>form" action="<?php echo $g_options['scripturl']; ?>?mode=admin&amp;game=<?php echo $gamecode; ?>&task=<?php echo $code; ?>#<?php echo $code; ?>">
 				<div class="table-responsive ms-4">
 					<table class="table">